Found some failing tests
